requirements.txt
coverage==7.3.2
execnet==2.0.2
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.4.0
python-dateutil==2.8.2
six==1.16.0